Move InjectionContext to module.rs and rename to InjectionModule
